Rename building.py to avoid confusions
