+++ default/scripting/buildings/buildings_macros.py
